requirements.txt
openai
flask
sqlalchemy
flask-sqlalchemy
flask-smorest
marshmallow
flask-migrate
python-dotenv
gunicorn
tenacity